Moved select inputs to the form section of components
