+++ src/js/design-system.js
@@ -3,6 +3,6 @@ import './main.js'
 
 //JS files for documnetations
 import 'components/_Progress-Bar/progress-bar-initialization.js'
-import 'components/_form-select-input/select-input-initialization.js'
+import 'components/forms/_select-input/select-input-initialization.js'
 import 'components/_expanding-components/accordion-initialization.js'
 import 'components/_expanding-components/section-initialization.js'
