+++ src/scss/main.scss
@@ -19,7 +19,7 @@
 @import 'components/forms/input-fields/input-field.scss';
 @import 'components/forms/text-areas/text-area.scss';
 @import 'components/forms/_checkboxes/checkbox.scss';
-@import 'components/_form-select-input/select-input.scss';
+@import 'components/forms/_select-input/select-input.scss';
 @import 'components/_buttons/button/buttons.scss';
 @import 'components/_buttons/button-floating/buttons-floating.scss';
 @import 'components/alert/alert.scss';
